refactor: move FeeDecorator methods from fee_utils.go to fee.go
